requirements.txt
pyTelegramBotAPI
wikipediaapi
python-dotenv